% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis.CD.R
\name{dis.CD}
\alias{dis.CD}
\title{Compute posterior mean based on discrete probability mass}
\usage{
dis.CD(x, prior.mass, h = 1)
}
\arguments{
\item{x}{Input n-dimensional noisy sparse Gaussian Sequence.}

\item{prior.mass}{Discrete n-dimensional prior mass vector. Each entry represents 1/n point mass at that entry.}

\item{h}{Error variance. Default to be 1. In real applications we assume it could be reliably estimated.}
}
\value{
Posterior mean estimator of the high dimensional Gaussian sequence.
}
\description{
This function is aimed to compute posterior mean vector based on discrete prior mass estimated by
sparseVBDP function.
}
\examples{
truemu=c(rep(0,900), rep(2,90),rep(10,10))
x=rnorm(1000, truemu)
results=sparseVBDP(x,1,6,0.01)
prior=results$prior
mu=dis.CD(x,prior)
}
\keyword{VBDP}

