% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPclassifier.R
\name{DPclassifier}
\alias{DPclassifier}
\title{Dirichlet Process based High Dimensional Empirical Bayes Linear stimator}
\usage{
DPclassifier(Train, Test, y, alpha, sigma, w, T0 = 10, nfolds = 10,
  combine = FALSE, sparse = TRUE)
}
\arguments{
\item{Train}{Matrix of training dataset.}

\item{Test}{Matrix of testing dataset.}

\item{y}{Factor of class labels of training set}

\item{alpha}{Concentration parameter of Dirichlet process.}

\item{sigma}{Standard deviation of Normal component in the base measure.}

\item{w}{Weight of zero component in the base measure.}

\item{T0}{Upper bound of number of clusters. Default to be 10.}

\item{nfolds}{Number of batches we have to apply sparseVBDP function.}

\item{combine}{If combine is TRUE, we merge the component with small means (<0.5) into 
zero component. Default to be FALSE.}

\item{sparse}{A binary indicator. If sparse is TRUE, we use dis.CD.sparse function to
compute a sparse version of Empirical Bayes mean difference estimator; otherwise we use dis.CD function 
to compute Empirical Bayes mean estimator.}
}
\value{
predicted class labels of observations in test dataset
}
\description{
This function is aimed to construct a linear classification rule based on multi-batch 
Dirichlet process empirical Bayes Classifier. Multi-batch 
Dirichlet process empirical Bayes is used to estimate 
mean difference to eliminate irrelavent features; then apply Naive Bayes classifier
}
\examples{
#Leukemia data set classification
set.seed(100)
data(leukemia)
Train=as.matrix(leukemia[1:38,-1]);
Test=as.matrix(leukemia[-(1:38),-1]);
trainlabel=leukemia[1:38,1];
testlabel=leukemia[-(1:38),1];
table(DPclassifier(Train,Test,trainlabel,1,4,0.995),testlabel)
}
\keyword{VBDP}

