% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisparseVBDP.R
\name{multisparseVBDP}
\alias{multisparseVBDP}
\title{multi-batch Dirichlet Process based Empirical Bayes Estimator}
\usage{
multisparseVBDP(x, alpha, sigma, w, T0 = 10, nfolds = 10, combine = FALSE)
}
\arguments{
\item{x}{Input noisy sparse Gaussian Sequence.}

\item{alpha}{Concentration parameter of Dirichlet Process.}

\item{sigma}{Standard deviation of Normal component in the base measure.}

\item{w}{Weight of zero component in the base measure.}

\item{T0}{Upper bound of number of clusters. Default to be 10.}

\item{nfolds}{Number of batches we have to apply sparseVBDP function.}

\item{combine}{If combine is TRUE, we merge the component with small means (<0.5) into 
zero component. Default to be FALSE.}
}
\value{
\emph{prior} Discrete prior vector containing each entry's estimated prior mean.

\emph{csize} The number of unique values in \emph{prior}.

\emph{prob} Probability of each element of unknown mean vector to be zero.
}
\description{
This function is aimed to estimate mean vector of sparse Gaussian sequence in the extremely
sparse case. Because of "rich gets richer" phenomenon, Simply using sparseVBDP function in extremely sparse case will
result in an estimator of all zero. To revise this estimator, we randomly
divide the whole vector into several folds and then apply sparseVBDP function. The 
revised estimator is used to construct high dimensional linear classification rule.
}
\examples{
truemu=c(rep(0,900), rep(2,90),rep(10,10))
x=rnorm(1000, truemu)
results=multisparseVBDP(x,1,6,0.01)
prior=results$prior
}
\keyword{VBDP}

