% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseVBDP.R
\name{sparseVBDP}
\alias{sparseVBDP}
\title{Sparse Dirichlet Process based Empirical Bayes Estimator using Variational Bayes Algorithm}
\usage{
sparseVBDP(x, alpha, sigma, w, T0 = 10, combine = FALSE)
}
\arguments{
\item{x}{Input noisy sparse Gaussian Sequence.}

\item{alpha}{Concentration parameter of Dirichlet Process.}

\item{sigma}{Standard deviation of Normal component in the base measure.}

\item{w}{Weight of zero component in the base measure.}

\item{T0}{Upper bound of number of clusters. Default to be 10.}

\item{combine}{If combine is TRUE, we merge the component with small means (<0.5) into 
zero component. Default to be FALSE.}
}
\value{
\emph{prior} Discrete prior vector containing each entry's estimated prior mean.

\emph{csize} The number of unique values in \emph{prior}.

\emph{prob} Probability of each element of unknown mean vector to be zero.
}
\description{
This function is aimed to estimate means of sparse Gaussian sequence.
}
\examples{
truemu=c(rep(0,900), rep(2,90),rep(10,10))
x=rnorm(1000, truemu)
results=sparseVBDP(x,1,6,0.01)
prior=results$prior
}
\keyword{VBDP}

